safari-534.53-branch/LayoutTests/fast/profiler/constructor.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.setJavaScriptProfilingEnabled(true);
}

function startTest()
{
    console.profile("Using a constructor.");

    var myObject = new fakeObject(1, 2);

    endTest();
}

function fakeObject (x, y)
{
    var myArray = new Array()
    myArray[0] = 5;
}
</script>
</head>

<body onload="startTest()">
This page's JavaScript has a call to new() in it.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be a call to fakeObject() and Array()
in it.
<div id="output"></div>
</body>
</html>
